Spin off TaskPacketDatabase into its own class
