+++ src/web/server/jaia_portal.py
@@ -15,6 +15,8 @@ from jaiabot.messages.engineering_pb2 import Engineering
 from jaiabot.messages.jaia_dccl_pb2 import *
 from jaiabot.messages.hub_pb2 import HubStatus
 
+from task_packet_database import *
+
 import google.protobuf.json_format
 
 from time import sleep
@@ -23,6 +25,7 @@ from pprint import *
 from typing import *
 from datetime import *
 from math import *
+from utils import *
 
 import logging
 
@@ -30,69 +33,9 @@ import logging
 BOT_PATH_UTIME_THRESHOLD = 2_000_000
 
 
-def now():
-    return int(datetime.now().timestamp() * 1e6)
-
-
-def utcnow():
-    return int(datetime.utcnow().timestamp() * 1e6)
-
-def utime(d: datetime):
-    '''Returns the utime for a datetime object'''
-    return int(d.replace(tzinfo=timezone.utc).timestamp() * 1e6)
-
-def floatFrom(obj):
-    try:
-        return float(obj)
-    except:
-        return None
-
-
 def protobufMessageToDict(message):
     return google.protobuf.json_format.MessageToDict(message, preserving_proto_field_name=True)
 
-
-def filterDuplicateTaskPackets(taskPackets: List[dict]):
-    """Filters duplicate task packets that can occur after data offloading. This works
-    by indexing the task packets by a (bot_id, reduced_time) tuple and checking neighboring
-    reduced_time values for duplicates.
-
-    Args:
-        taskPackets (List[dict]): Unfiltered list of task packets.
-    Returns:
-        (List[dict]): Filtered list of task packets.
-    """
-    # Maps (bot_id, reduced_time) to TaskPacket
-    taskPacketLookup: Dict[tuple, dict] = {}
-
-    for taskPacket in taskPackets:
-        bot_id = taskPacket['bot_id']
-        reducedStartTime = reduceTime(taskPacket['start_time'])
-
-        # Check neighboring bins as well for task packets, just in case start_time was on
-        # the cusp of being rounded up/down
-        if (bot_id, reducedStartTime) in taskPacketLookup or \
-           (bot_id, reducedStartTime - 1) in taskPacketLookup or \
-           (bot_id, reducedStartTime + 1) in taskPacketLookup:
-            continue
-        else:
-            taskPacketLookup[(bot_id, reducedStartTime)] = taskPacket
-        
-    return list(taskPacketLookup.values())
-
-def reduceTime(time: int):
-        """Does integer division to give the floored Unix timestamp in seconds.
-
-        Args:
-            time (int): Unix timestamp in microseconds.
-
-        Returns:
-            int: Unix timestamp in seconds, rounded down.
-        """
-        # This BIN_LENGTH can be adjusted if desired, but DCCL time2 codec rounds to the nearest 
-        # second, (1 million microseconds)
-        BIN_LENGTH = 1_000_000
-        return int(time) // BIN_LENGTH
 
 
 class BotPathPoint(NamedTuple):
@@ -123,17 +66,8 @@ class Interface:
     # MetaData
     metadata = {}
 
-    all_task_packets: List[Dict] = []
-    task_packets_version = 0
-    offloaded_task_packet_files_prev = -1
-    offloaded_task_packet_files_curr = 0
-    taskPacketPath = '/var/log/jaiabot/bot_offload/'
-
-    # Set the initial time for checking for task packet files
-    start_task_packet_check_time = now()
-
-    # Time between checking for task packet files (10 Seconds)
-    task_packet_check_interval = 10_000_000
+    # Task packet database
+    task_packet_database = TaskPacketDatabase()
 
     def __init__(self, goby_host=('localhost', 40000), read_only=False):
         self.goby_host = goby_host
@@ -182,12 +116,7 @@ class Interface:
                 data = self.sock.recv(10000)
                 self.process_portal_to_client_message(data)
 
-                # Check if the desired time interval has passed
-                if now() - self.start_task_packet_check_time >= self.task_packet_check_interval:
-                    self.load_taskpacket_files()
-                    
-                    # Reset the start time
-                    self.start_task_packet_check_time = now()
+                self.task_packet_database.loop()
 
             except socket.timeout:
                 self.ping_portal()
@@ -528,8 +457,7 @@ class Interface:
 
     def process_task_packet(self, task_packet_message):
         task_packet = protobufMessageToDict(task_packet_message)
-        self.all_task_packets.append(task_packet)
-        self.task_packets_version += 1
+        self.task_packet_database.add_task_packet(task_packet)
 
     def process_active_mission_plan(self, bot_id, active_mission_plan):
         try:
@@ -538,41 +466,6 @@ class Interface:
         except IndexError:
             logging.warning(f'Received active mission plan for unknown bot {active_mission_plan.bot_id}')
 
-    def get_task_packets_subset(self, start_date, end_date):
-        """Selects TaskPackets between the provided date bounds
-        Args:
-            start_date (str): Provides the lower bound
-            end_date (str): Provides the upper bound
-        Returns:
-            list: Subset of TaskPackets between specified dates
-        """
-        start_index = bisect.bisect_left(
-            list(map(lambda task_packet: int(task_packet['start_time']),  self.all_task_packets)), 
-            utime(start_date)
-        )
-
-        if end_date == "":
-            end_index = len(self.all_task_packets)
-        else:
-            end_index = bisect.bisect_right(
-                list(map(lambda task_packet: int(task_packet['start_time']),  self.all_task_packets)),
-                utime(end_date)
-            )
-
-        return self.all_task_packets[start_index: end_index]
-
-    def get_task_packets(self, start_date, end_date):
-        if start_date is None or end_date is None:
-            return []
-
-        return self.get_task_packets_subset(start_date, end_date)
-    
-    def get_task_packets_version(self) -> int:
-        """Gets the version of the set of TaskPackets.  When task packet(s) gets added, removed, changed, etc., the version gets incremented.
-        Returns:
-            int: The version of the set of TaskPackets
-        """
-        return self.task_packets_version
 
     # Contour map
     
@@ -586,12 +479,12 @@ class Interface:
         Returns:
             dict[str, any]: A GeoJSON dictionary representing a depth color map for the bottom dives.
         """
-        return pyjaia.contours.taskPacketsToColorMap(self.get_task_packets(start_date, end_date))
+        return pyjaia.contours.taskPacketsToColorMap(self.task_packet_database.get_task_packets(start_date, end_date))
 
     # Drift map
 
     def get_drift_map(self, start_date, end_date):
-        return pyjaia.drift_interpolation.taskPacketsToDriftMarkersGeoJSON(self.get_task_packets(start_date, end_date))
+        return pyjaia.drift_interpolation.taskPacketsToDriftMarkersGeoJSON(self.task_packet_database.get_task_packets(start_date, end_date))
 
     # Bot paths
 
@@ -614,30 +507,3 @@ class Interface:
 
     def get_Metadata(self):
         return self.metadata
-
-    def load_taskpacket_files(self):
-        """Appends TaskPackets from *.taskpacket files in the bot_offload directory 
-           to the list of all TaskPackets. Removes duplicates between offloaded and live
-           TaskPackets and sorts the list by start time.
-        Returns: None
-        """
-        self.offloaded_task_packet_file_curr = len(glob.glob(self.taskPacketPath + '*.taskpacket'))
-
-        if self.offloaded_task_packet_file_curr != self.offloaded_task_packet_files_prev:
-            self.offloaded_task_packet_files_prev = self.offloaded_task_packet_file_curr
-        else:
-            return
-
-        for filePath in glob.glob(self.taskPacketPath + '*.taskpacket'):
-            filePath = Path(filePath)
-
-            for line in open(filePath):
-                try:
-                    taskPacket: Dict = json.loads(line)
-                    self.all_task_packets.append(taskPacket)
-                    self.task_packets_version += 1
-                except json.JSONDecodeError as e:
-                    logging.warning(f"Error decoding JSON line: {line} because {e}")
-
-        self.all_task_packets = filterDuplicateTaskPackets(self.all_task_packets)
-        self.all_task_packets.sort(key=lambda taskPacket: int(taskPacket['start_time']))
